requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-smart-selects==1.6.0
PyMySQL==1.0.2
pysqlite3-binary==0.5.0
sqlparse==0.4.3
whitenoise==6.3.0
